Create a simple cache
This will prevent the need from retrieving the value for each and every
call. This may not be overly important for env vars that are currently
implemented. However, this could become a big cost saver for when you
are retrieving your values from an external service.

We also renamed the MyConfig.validate!/0 function to MyConfig.load!/0.
You are now required to call `MyConfig.load!/0` before you try using
your configuration as that is what will actually grab the values and
store them into the cache.