Create JsonEndpoint Provider
This is meant to enable the user to retrieve values from a JSON
endpoint.

This also includes a way to make value names to how we want to reference
them in our own systems. This means that if our json endpoint (or any
other provider) gives us the key `foo-bar`, we can still use `foo_bar` in
our own code. Please see `opt_7` in `json_endpoint_test.exs` for an
example of how to do this. This should work for all providers, and not
just for the JsonEndpoint.

If the JsonEndpoint provider is unable to make the call to the specified
endpoint, it will log a warning and pass nil values back to the caller
and will still succeed assuming you have created default values for all
params. This is mainly useful in development where you may not actually
have another HTTP server running in order to retrieve values.