Remove cache polymorphism
The polymorphism did not really add anything to the implementation and
made it messier than it needed to be.

This change also turned the config module into a GenServer that will
start a named ETS table (the cache) and load all of the data from the
source into the cache for quick retrieval. Currently, the only purpose
of the GenServer is to own the ETS table.