Fix credo warnings
When writing the spec for Provider.Cache.get/2, I realized that the
error case is not something that should ever happen anymore. One of two
things should happen when calling that function. Either,

* The user forgot to start the GenServer, in which case ets will
complains that the table does not exist; or
* The user did start the GenServer, in which case the ets table exists
and it will be initialized with the data from the configured source.

And if you happen to not give a default to a given param and forget to
configure it, the GenServer will fail to start because of it.