Support ecto config
Ecto has mix tasks that will attempt to start the repo. In some of these
scenarios, the application will not yet have been started. Most notable
are ecto.create and ecto.migrate. These are aliased for you in a brand
new phoenix project under `mix test`.

With this new `Provider.ecto_config/3` function, we ensure that the
given config process is running, and we will also shut it down
afterwards to ensure that when your application starts, we do not have
an issue with the process already being started.